tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.size.attributes.removed.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.size.attributes.removed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.size.attributes.removed</h1>
<p class="desc">Removing content attributes reverts to default size</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="120" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Removing content attributes reverts to default size");
_addTest(function(canvas, ctx) {

  _assertSame(canvas.width, 120, "canvas.width", "120");
  canvas.removeAttribute('width');
  _assertSame(canvas.width, 300, "canvas.width", "300");

});
</script>
